# Tweak object instantiation hot paths