--- a/test/Spring/Spring.Benchmark/Spring.Benchmark.csproj
+++ b/test/Spring/Spring.Benchmark/Spring.Benchmark.csproj
@@ -7,7 +7,7 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="BenchmarkDotNet" Version="0.11.1" />
+      <PackageReference Include="BenchmarkDotNet" Version="0.11.2" />
     </ItemGroup>
 
     <ItemGroup>