… src/Spring/Spring.Core/Objects/Factory/Support/ConstructorResolver.cs
… src/Spring/Spring.Core/Objects/Factory/Support/ObjectDefinitionValueResolver.cs
… src/Spring/Spring.Core/Objects/ObjectWrapper.cs