## Old 


``` ini

BenchmarkDotNet=v0.11.1, OS=Windows 10.0.17763
AMD Ryzen 7 2700X Eight-Core Processor (Max: 4.00GHz), 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=2.1.403
  [Host] : .NET Core 2.1.5 (CoreCLR 4.6.26919.02, CoreFX 4.6.26919.02), 64bit RyuJIT
  Clr    : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.3190.0
  Core   : .NET Core 2.1.5 (CoreCLR 4.6.26919.02, CoreFX 4.6.26919.02), 64bit RyuJIT


```
|           Method |  Job | Runtime | Iterations |      Mean |     Error |    StdDev |     Gen 0 |  Allocated |
|----------------- |----- |-------- |----------- |----------:|----------:|----------:|----------:|-----------:|
| ResolveTransient |  Clr |     Clr |       5000 | 26.579 ms | 0.1103 ms | 0.1032 ms | 1968.7500 |  1525.1 KB |
| ResolveSingleton |  Clr |     Clr |       5000 |  4.624 ms | 0.0279 ms | 0.0261 ms |  453.1250 |  351.96 KB |
| ResolveTransient | Core |    Core |       5000 | 25.780 ms | 0.1411 ms | 0.1320 ms | 1968.7500 | 1524.94 KB |
| ResolveSingleton | Core |    Core |       5000 |  4.404 ms | 0.0718 ms | 0.0600 ms |  453.1250 |  351.91 KB |


## New

``` ini

BenchmarkDotNet=v0.11.2, OS=Windows 10.0.17763.107 (1809/October2018Update/Redstone5)
AMD Ryzen 7 2700X, 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=2.1.403
  [Host] : .NET Core 2.1.5 (CoreCLR 4.6.26919.02, CoreFX 4.6.26919.02), 64bit RyuJIT
  Clr    : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.3190.0
  Core   : .NET Core 2.1.5 (CoreCLR 4.6.26919.02, CoreFX 4.6.26919.02), 64bit RyuJIT


```
|           Method |  Job | Runtime | Iterations |      Mean |     Error |    StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|----------------- |----- |-------- |----------- |----------:|----------:|----------:|------------:|------------:|------------:|--------------------:|
| ResolveTransient |  Clr |     Clr |       5000 | 22.868 ms | 0.0957 ms | 0.0848 ms |   1968.7500 |           - |           - |           1525.1 KB |
| ResolveSingleton |  Clr |     Clr |       5000 |  4.634 ms | 0.0123 ms | 0.0115 ms |    453.1250 |           - |           - |           351.96 KB |
| ResolveTransient | Core |    Core |       5000 | 22.593 ms | 0.1102 ms | 0.1031 ms |   1968.7500 |           - |           - |          1524.94 KB |
| ResolveSingleton | Core |    Core |       5000 |  4.419 ms | 0.0528 ms | 0.0494 ms |    453.1250 |           - |           - |           351.91 KB |